Created ContactBookContext
Combined all previous Context Classes into ContactBookContext
