--- CustomerContactBook/Models/CustomerContext.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace CustomerContactBook.Models
-{
-    public class CustomerContext : DbContext
-    {
-        public CustomerContext(DbContextOptions<CustomerContext> options)
-            : base(options)
-        {
-
-        }
-
-        public DbSet<Customer> Customers { get; set; }
-    }
-}
