+++ CustomerContactBook/Models/CustomerGroup.cs
@@ -1,4 +1,6 @@
-﻿namespace CustomerContactBook.Models
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace CustomerContactBook.Models
 {
     /// <summary>
     /// Model Class for Customer Group
@@ -9,6 +11,7 @@
         /// identifier for CustomerGroup
         /// </summary>
         /// <example>7032</example>
+        [Key]
         public long Id { get; set; }
         /// <summary>
         /// Group Name
