+++ CustomerContactBook/Program.cs
@@ -4,12 +4,8 @@ using CustomerContactBook.Models;
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
-builder.Services.AddDbContext<CustomerContext>(opt =>
-               opt.UseInMemoryDatabase("CustomerList"));
-builder.Services.AddDbContext<CustomerGroupContext>(opt =>
-               opt.UseInMemoryDatabase("GroupList"));
-builder.Services.AddDbContext<GroupMemberContext>(opt =>
-               opt.UseInMemoryDatabase("GroupMemberList"));
+builder.Services.AddDbContext<ContactBookContext>(opt =>
+               opt.UseInMemoryDatabase("ContactBook"));
 builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
